aws unit tests done
